@@ -4,7 +4,9 @@
   "private": true,
   "scripts": {
     "lint": "xo",
-    "test": "mocha --compilers js:babel-core/register --require tests/tools/requireCSS.js tests/**/*.spec.js{,x}",
+    "test": "npm run test:coverage --silent",
+    "test:mocha": "mocha --compilers js:babel-core/register --require tests/tools/requireCSS.js tests/**/*.spec.js{,x}",
+    "test:coverage": "nyc --all --include=src --extension .jsx --reporter=lcov --reporter=text-summary --report-dir=coverage npm run test:mocha",
     "build": "gulp build",
     "production": "NODE_ENV=production gulp build",
     "dev": "gulp dev",
@@ -21,11 +23,11 @@
     "debug": "^2.6.0",
     "express": "^4.14.0",
     "morgan": "^1.7.0",
-    "pug": "^2.0.0-beta11",
+    "pug": "^2.0.0-rc.2",
     "serve-favicon": "^2.3.2"
   },
   "devDependencies": {
-    "autoprefixer": "^6.6.1",
+    "autoprefixer": "^7.1.0",
     "babel-plugin-lodash": "^3.2.11",
     "babel-plugin-react-intl": "^2.3.0",
     "babel-preset-es2015": "^6.18.0",
@@ -35,14 +37,14 @@
     "babelify": "^7.3.0",
     "browser-sync": "^2.18.6",
     "browserify": "^14.3.0",
-    "chai": "^3.5.0",
+    "chai": "^4.0.0",
     "change-case": "^3.0.0",
     "classnames": "^2.2.5",
-    "cssnano": "^3.10.0",
-    "del": "^2.2.2",
-    "enzyme": "^2.7.1",
-    "eslint-config-xo-react": "^0.11.1",
-    "eslint-plugin-react": "^6.9.0",
+    "cssnano": "^4.0.0-rc.1",
+    "del": "^3.0.0",
+    "enzyme": "^2.9.1",
+    "eslint-config-xo-react": "^0.12.0",
+    "eslint-plugin-react": "^7.1.0",
     "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
     "gulp-imagemin": "^3.1.1",
@@ -50,34 +52,37 @@
     "gulp-rev-all": "^0.9.7",
     "gulp-sourcemaps": "^2.3.1",
     "gulp-to-ico": "^1.0.0",
-    "gulp-uglify": "^2.0.0",
+    "gulp-uglify": "^3.0.0",
     "gulp-util": "^3.0.8",
-    "gw2api-client": "^5.0.0",
-    "gw2api-extension": "file:lib/gw2api-extension",
-    "gw2be-api-extension-data": "file:lib/gw2be-api-extension-data",
-    "gw2be-build-string": "file:lib/gw2be-build-string",
+    "gw2api-client": "^5.6.2",
+    "gw2api-extension": "file:./lib/gw2api-extension",
+    "gw2be-api-extension-data": "file:./lib/gw2be-api-extension-data",
+    "gw2be-build-string": "file:./lib/gw2be-build-string",
     "lodash": "^4.17.4",
     "mocha": "^3.2.0",
-    "modular-css": "^3.0.0",
-    "postcss-import": "^9.1.0",
+    "modular-css": "^6.0.0",
+    "nyc": "^11.1.0",
+    "postcss-import": "^10.0.0",
     "postcss-urlrewrite": "^0.2.1",
-    "react": "^15.4.2",
-    "react-addons-css-transition-group": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-addons-transition-group": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-addons-css-transition-group": "^15.6.0",
+    "react-addons-test-utils": "^15.6.0",
+    "react-addons-transition-group": "^15.6.0",
+    "react-dom": "^15.6.1",
     "react-intl": "^2.2.2",
-    "react-onclickoutside": "^5.8.3",
+    "react-onclickoutside": "^6.1.1",
     "react-redux": "^5.0.2",
-    "redux": "^3.6.0",
+    "redux": "^3.7.1",
     "redux-actions": "^2.0.1",
     "redux-promise": "^0.5.3",
     "redux-sync-reducer": "^0.0.4",
     "redux-thunk": "^2.1.0",
     "reselect": "^3.0.0",
-    "run-sequence": "^1.2.2",
+    "run-sequence": "^2.0.0",
     "serviceworker-storage": "^0.1.0",
-    "typeface-open-sans": "0.0.22",
+    "through2": "^2.0.3",
+    "typeface-open-sans": "0.0.29",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
     "watchify": "^3.8.0",
