@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import SkillTooltip from '~/components/Tooltips/Skills/TooltipContainer';
 import SkillIcon from '../../Icon';
@@ -9,37 +10,35 @@ class WeaponBar extends Component {
         this.renderSkill = this.renderSkill.bind(this);
     }
 
-    getSkillInSlot({ mainhand, offhand, isTwoHanded, weapons, skills, attunement }) {
+    getSkillInSlot({ activeMainhandWeaponId, activeOffhandWeaponId, isTwoHanded, availableWeaponObjects, availableSkillObjects, activeAttunement }) {
         return index => {
             const getFromMainhand = index < 3 || isTwoHanded;
 
             // handle empty weapon
-            if ((getFromMainhand && !mainhand) || (!getFromMainhand && !offhand)) {
+            if ((getFromMainhand && !activeMainhandWeaponId) || (!getFromMainhand && !activeOffhandWeaponId)) {
                 return null;
             }
 
-            const weapon = getFromMainhand ? mainhand : offhand;
+            const weapon = getFromMainhand ? activeMainhandWeaponId : activeOffhandWeaponId;
             const slotName = `Weapon_${index + 1}`;
 
             // find all matching skills for the requested slot
-            const skillsForSlot = weapons[weapon].skills
+            const skillsForSlot = availableWeaponObjects[weapon].skills
                 // filter slot
                 .filter(skill => skill.slot === slotName)
                 // filter offhand (thief)
-                .filter(skill => !skill.offhand || skill.offhand === offhand || (skill.offhand === 'Nothing' && !offhand))
+                .filter(skill => !skill.offhand || skill.offhand === activeOffhandWeaponId || (skill.offhand === 'Nothing' && !activeOffhandWeaponId))
                 // filter attunement (elementalist)
-                .filter(skill => !skill.attunement || skill.attunement === attunement);
+                .filter(skill => !skill.attunement || skill.attunement === activeAttunement);
 
             // return the skill object
-            return skillsForSlot[0] && skills[skillsForSlot[0].id];
+            return skillsForSlot[0] && availableSkillObjects[skillsForSlot[0].id];
         };
     }
 
     render() {
         const getSkillInSlot = this.getSkillInSlot(this.props);
-
         const slots = [0, 1, 2, 3, 4].map(getSkillInSlot);
-
         return (
             <div>
                 {slots.map(this.renderSkill)}
@@ -63,20 +62,22 @@ class WeaponBar extends Component {
 }
 
 WeaponBar.propTypes = {
-    mainhand: React.PropTypes.string,
-    offhand: React.PropTypes.string,
-    isTwoHanded: React.PropTypes.bool,
-    attunement: React.PropTypes.string,
-    weapons: React.PropTypes.objectOf(React.PropTypes.shape({
-        skills: React.PropTypes.arrayOf(React.PropTypes.shape({
-            id: React.PropTypes.number.isRequired,
-            slot: React.PropTypes.string.isRequired,
-            attunement: React.PropTypes.string,
-            offhand: React.PropTypes.string
+    isTwoHanded: PropTypes.bool,
+
+    // Redux states
+    activeMainhandWeaponId: PropTypes.string,
+    activeOffhandWeaponId: PropTypes.string,
+    activeAttunement: PropTypes.string,
+    availableSkillObjects: PropTypes.object.isRequired,
+    availableWeaponObjects: PropTypes.objectOf(PropTypes.shape({
+        skills: PropTypes.arrayOf(PropTypes.shape({
+            id: PropTypes.number.isRequired,
+            slot: PropTypes.string.isRequired,
+            attunement: PropTypes.string,
+            offhand: PropTypes.string
         })).isRequired,
-        specialization: React.PropTypes.number
+        specialization: PropTypes.number
     })).isRequired,
-    skills: React.PropTypes.object.isRequired
 };
 
 export default WeaponBar;
