@@ -1,40 +1,45 @@
 import map from 'lodash/map';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { Select } from '~/components/Inputs';
 
 // a weapon is available if it doesn't require a specialization
 // or if the specialization is active.
-const isWeaponAvailable = (weapon, activeSpecializations) =>
-    !weapon.specialization || activeSpecializations.some(id => id === weapon.specialization);
+const isWeaponAvailable = (weapon, selectedSpecializationIds) =>
+    !weapon.specialization || selectedSpecializationIds.some(id => id === weapon.specialization);
 
-const renderSelect = (weapons, value, onChange, activeSpecializations) => (
+const renderSelect = (weapons, value, onChange, selectedSpecializationIds) => (
     <Select value={value} onChange={onChange} placeholder="None">
         {map(weapons, (weapon, name) => (
-            <option key={name} value={name} disabled={!isWeaponAvailable(weapon, activeSpecializations)}>{name}</option>
+            <option key={name} value={name} disabled={!isWeaponAvailable(weapon, selectedSpecializationIds)}>{name}</option>
         ))}
     </Select>
 );
 
 const WeaponSelection = ({
-    mainhandWeapons, onMainhandChange, activeMainhand,
-    offhandWeapons, onOffhandChange, activeOffhand,
-    twoHanded, activeSpecializations
+    availableMainhandWeaponObjects, onMainhandChange, activeMainhandWeaponId,
+    availableOffhandWeaponObjects, onOffhandChange, activeOffhandWeaponId,
+    twoHanded, selectedSpecializationIds
 }) => (
     <div>
-        {renderSelect(mainhandWeapons, activeMainhand, onMainhandChange, activeSpecializations)}
-        {!twoHanded && renderSelect(offhandWeapons, activeOffhand, onOffhandChange, activeSpecializations)}
+        {renderSelect(availableMainhandWeaponObjects, activeMainhandWeaponId, onMainhandChange, selectedSpecializationIds)}
+        {!twoHanded && renderSelect(availableOffhandWeaponObjects, activeOffhandWeaponId, onOffhandChange, selectedSpecializationIds)}
     </div>
 );
 
 WeaponSelection.propTypes = {
-    activeSpecializations: React.PropTypes.arrayOf(React.PropTypes.number),
-    mainhandWeapons: React.PropTypes.object,
-    offhandWeapons: React.PropTypes.object,
-    onMainhandChange: React.PropTypes.func.isRequired,
-    onOffhandChange: React.PropTypes.func.isRequired,
-    activeMainhand: React.PropTypes.string,
-    activeOffhand: React.PropTypes.string,
-    twoHanded: React.PropTypes.bool.isRequired
+    twoHanded: PropTypes.bool.isRequired,
+
+    // Events
+    onMainhandChange: PropTypes.func.isRequired,
+    onOffhandChange: PropTypes.func.isRequired,
+
+    // Redux states
+    activeMainhandWeaponId: PropTypes.string,
+    activeOffhandWeaponId: PropTypes.string,
+    availableMainhandWeaponObjects: PropTypes.object,
+    availableOffhandWeaponObjects: PropTypes.object,
+    selectedSpecializationIds: PropTypes.arrayOf(PropTypes.number)
 };
 
 export default WeaponSelection;
