@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import SpecializationIcon from './SpecializationIcon';
 import style from './selectionPopup.css';
@@ -10,7 +11,7 @@ class SelectionPopup extends React.Component {
 
     handleOnIconClick(id) {
         if (this.props.onSpecializationChange) {
-            this.props.onSpecializationChange(id, this.props.activeSpecializationIds, this.props.specializations);
+            this.props.onSpecializationChange(id, this.props.selectedSpecializationIds, this.props.availableSpecializationObjects);
             if (this.props.onWantsClose) {
                 // There is a small bug that causes the specialization to not update when the popup isn't closed.
                 // Now we want to close it after selecting a specialization anyway, so it doesn't matter much.
@@ -23,13 +24,13 @@ class SelectionPopup extends React.Component {
     renderSpecializations(ids) {
         if (ids) {
             return ids.map(id => {
-                const specialization = this.props.specializations[id];
+                const specialization = this.props.availableSpecializationObjects[id];
                 return (
                     <SpecializationIcon
                         key={id}
                         icon={specialization.icon}
                         id={id}
-                        isSelected={this.props.activeSpecializationIds.indexOf(id) > -1}
+                        isSelected={this.props.selectedSpecializationIds.indexOf(id) > -1}
                         name={specialization.name}
                         onClick={this.handleOnIconClick}/>
                 );
@@ -66,17 +67,17 @@ class SelectionPopup extends React.Component {
 }
 
 SelectionPopup.propTypes = {
-    supportsElite: React.PropTypes.bool,
+    supportsElite: PropTypes.bool,
 
     // Events
-    onSpecializationChange: React.PropTypes.func,
-    onWantsClose: React.PropTypes.func,
+    onSpecializationChange: PropTypes.func,
+    onWantsClose: PropTypes.func,
 
     // Redux states
-    activeSpecializationIds: React.PropTypes.arrayOf(React.PropTypes.number),
-    availableCoreSpecializationIds: React.PropTypes.arrayOf(React.PropTypes.number),
-    availableEliteSpecializationIds: React.PropTypes.arrayOf(React.PropTypes.number),
-    specializations: React.PropTypes.object
+    selectedSpecializationIds: PropTypes.arrayOf(PropTypes.number),
+    availableCoreSpecializationIds: PropTypes.arrayOf(PropTypes.number),
+    availableEliteSpecializationIds: PropTypes.arrayOf(PropTypes.number),
+    availableSpecializationObjects: PropTypes.object
 };
 
 export default SelectionPopup;
