@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import TraitTooltip from '~/components/Tooltips/Traits/TooltipContainer';
 import TraitConnection from '../TraitConnection';
@@ -14,8 +15,8 @@ class TraitTier extends React.Component {
     }
 
     getIcon(id) {
-        if (id && this.props.traits && this.props.traits[id]) {
-            return this.props.traits[id].icon;
+        if (id && this.props.availableTraitObjects && this.props.availableTraitObjects[id]) {
+            return this.props.availableTraitObjects[id].icon;
         }
     }
 
@@ -41,8 +42,8 @@ class TraitTier extends React.Component {
             line = lineTranslate[this.props.majorTraitIds.indexOf(this.props.selectedMajorTraitId)];
         }
 
-        const minorTrait = this.props.traits && this.props.traits[this.props.minorTraitId];
-        const majorTraits = this.props.traits ? this.props.majorTraitIds.map(id => this.props.traits[id]) : [null, null, null];
+        const minorTrait = this.props.availableTraitObjects && this.props.availableTraitObjects[this.props.minorTraitId];
+        const majorTraits = this.props.availableTraitObjects ? this.props.majorTraitIds.map(id => this.props.availableTraitObjects[id]) : [null, null, null];
 
         return (
             <div className={style.traitTier} onClick={this.handleClick}>
@@ -81,17 +82,17 @@ class TraitTier extends React.Component {
 }
 
 TraitTier.propTypes = {
-    traitTier: React.PropTypes.number,
+    traitTier: PropTypes.number,
 
     // Events
-    onBackgroundClick: React.PropTypes.func,
-    onTraitChange: React.PropTypes.func,
+    onBackgroundClick: PropTypes.func,
+    onTraitChange: PropTypes.func,
 
     // Redux states
-    majorTraitIds: React.PropTypes.arrayOf(React.PropTypes.number),
-    minorTraitId: React.PropTypes.number,
-    selectedMajorTraitId: React.PropTypes.number,
-    traits: React.PropTypes.object
+    availableTraitObjects: PropTypes.object,
+    majorTraitIds: PropTypes.arrayOf(PropTypes.number),
+    minorTraitId: PropTypes.number,
+    selectedMajorTraitId: PropTypes.number,
 };
 
 export default TraitTier;
