@@ -1,5 +1,6 @@
 export function serveFromCacheFallbackToNetwork(request, cacheName) {
-    return caches.match(request)
+    return caches.open(cacheName)
+        .then(cache => cache.match(request))
         .then(cached => cached || fetch(request).then(response => saveRequestToCache(request, response, cacheName))
     );
 }
@@ -7,10 +8,16 @@ export function serveFromCacheFallbackToNetwork(request, cacheName) {
 export function serveFromNetworkFallbackToCache(request, cacheName) {
     return fetch(request)
         .then(response => saveRequestToCache(request, response, cacheName))
-        .catch(() => caches.match(request));
+        .catch(
+            () => caches.open(cacheName).then(cache => cache.match(request))
+        );
 }
 
 export function saveRequestToCache(request, response, cacheName) {
+    if(!response.ok) {
+        return response;
+    }
+
     return caches.open(cacheName).then(cache => {
         cache.put(request, response.clone());
         return response;
