@@ -1,48 +1,48 @@
 import { createAction } from 'redux-actions';
 import { createApiAction, convertToIndexed } from './utils';
 
-export const SET_WEAPON_SET = 'SET_WEAPON_SET';
+export const SET_ACTIVE_WEAPON_SET = 'SET_ACTIVE_WEAPON_SET';
 
-export const SET_MAINHAND_WEAPON = 'SET_MAINHAND_WEAPON';
-export const SET_OFFHAND_WEAPON = 'SET_OFFHAND_WEAPON';
-export const WIPE_ALL_WEAPONS = 'WIPE_ALL_WEAPONS';
+export const SET_ACTIVE_MAINHAND_WEAPON_ID = 'SET_ACTIVE_MAINHAND_WEAPON_ID';
+export const SET_ACTIVE_OFFHAND_WEAPON_ID = 'SET_ACTIVE_OFFHAND_WEAPON_ID';
+export const WIPE_ALL_SELECTED_WEAPON_IDS = 'WIPE_ALL_SELECTED_WEAPON_IDS';
 
-export const SET_ATTUNEMENT = 'SET_ATTUNEMENT';
+export const SET_ACTIVE_ATTUNEMENT = 'SET_ACTIVE_ATTUNEMENT';
 
-export const FETCH_SKILLS = 'FETCH_SKILLS';
+export const FETCH_AVAILABLE_SKILLS = 'FETCH_AVAILABLE_SKILLS';
 
 /** Action to swap two set specializations with each other. Params: { specializationLine1, specializationLine2 } */
-export const setWeaponSet = createAction(SET_WEAPON_SET);
+export const setActiveWeaponSet = createAction(SET_ACTIVE_WEAPON_SET);
 
-export const setMainhandWeapon = createAction(SET_MAINHAND_WEAPON);
-export const setOffhandWeapon = createAction(SET_OFFHAND_WEAPON);
-export const wipeAllWeapons = createAction(WIPE_ALL_WEAPONS);
+export const setActiveMainhandWeaponId = createAction(SET_ACTIVE_MAINHAND_WEAPON_ID);
+export const setActiveOffhandWeaponId = createAction(SET_ACTIVE_OFFHAND_WEAPON_ID);
+export const wipeAllSelectedWeaponIds = createAction(WIPE_ALL_SELECTED_WEAPON_IDS);
 
-export const setAttunement = createAction(SET_ATTUNEMENT);
+export const setActiveAttunement = createAction(SET_ACTIVE_ATTUNEMENT);
 
-export const fetchSkills = createApiAction(
-    FETCH_SKILLS,
-    (state, api) => api.skills().many(state.skillIds).then(convertToIndexed)
+export const fetchAvailableSkills = createApiAction(
+    FETCH_AVAILABLE_SKILLS,
+    (state, api) => api.skills().many(state.availableSkillIds).then(convertToIndexed)
 );
 
 export default {
-    SET_WEAPON_SET,
+    SET_ACTIVE_WEAPON_SET,
 
-    SET_MAINHAND_WEAPON,
-    SET_OFFHAND_WEAPON,
-    WIPE_ALL_WEAPONS,
+    SET_ACTIVE_MAINHAND_WEAPON_ID,
+    SET_ACTIVE_OFFHAND_WEAPON_ID,
+    WIPE_ALL_SELECTED_WEAPON_IDS,
 
-    SET_ATTUNEMENT,
+    SET_ACTIVE_ATTUNEMENT,
 
-    FETCH_SKILLS,
+    FETCH_AVAILABLE_SKILLS,
 
-    setWeaponSet,
+    setActiveWeaponSet,
 
-    setMainhandWeapon,
-    setOffhandWeapon,
-    wipeAllWeapons,
+    setActiveMainhandWeaponId,
+    setActiveOffhandWeaponId,
+    wipeAllSelectedWeaponIds,
 
-    setAttunement,
+    setActiveAttunement,
 
-    fetchSkills
+    fetchAvailableSkills
 };
