@@ -1,34 +1,34 @@
 import { createAction } from 'redux-actions';
 import { createApiAction, convertToIndexed } from './utils';
 
-export const FETCH_TRAITS = 'FETCH_TRAITS';
-export const SET_MAJOR_TRAIT = 'SET_MAJOR_TRAIT';
-export const WIPE_ACTIVE_TRAITS = 'WIPE_ACTIVE_TRAITS';
-export const WIPE_ALL_ACTIVE_TRAITS = 'WIPE_ALL_ACTIVE_TRAITS';
+export const FETCH_AVAILABLE_TRAITS = 'FETCH_AVAILABLE_TRAITS';
+export const SET_SELECTED_MAJOR_TRAIT_ID = 'SET_SELECTED_MAJOR_TRAIT_ID';
+export const WIPE_SELECTED_TRAIT_IDS = 'WIPE_SELECTED_TRAIT_IDS';
+export const WIPE_ALL_SELECTED_TRAIT_IDS = 'WIPE_ALL_SELECTED_TRAIT_IDS';
 
 /** Action to fetch the current traits from the GW2 API. */
-export const fetchTraits = createApiAction(
-    FETCH_TRAITS,
-    (state, api) => api.traits().many(state.traitIds).then(convertToIndexed)
+export const fetchAvailableTraits = createApiAction(
+    FETCH_AVAILABLE_TRAITS,
+    (state, api) => api.traits().many(state.availableTraitIds).then(convertToIndexed)
 );
 
 /** Action to set a specific major trait in a trait tier of a specialization line. Params: { specializationLine, traitTier, traitId } */
-export const setMajorTrait = createAction(SET_MAJOR_TRAIT);
+export const setSelectedMajorTraitId = createAction(SET_SELECTED_MAJOR_TRAIT_ID);
 
-/** Action to wipe the active traits of a specific specialization line. Params: { specializationLine } */
-export const wipeActiveTraits = createAction(WIPE_ACTIVE_TRAITS);
+/** Action to wipe the selected traits of a specific specialization line. Params: { specializationLine } */
+export const wipeSelectedTraitIds = createAction(WIPE_SELECTED_TRAIT_IDS);
 
-/** Action to wipe all active traits of every specialization lines. Params: {} */
-export const wipeAllActiveTraits = createAction(WIPE_ALL_ACTIVE_TRAITS);
+/** Action to wipe all selected traits of all selected specialization lines. Params: {} */
+export const wipeAllSelectedTraitIds = createAction(WIPE_ALL_SELECTED_TRAIT_IDS);
 
 export default {
-    FETCH_TRAITS,
-    SET_MAJOR_TRAIT,
-    WIPE_ACTIVE_TRAITS,
-    WIPE_ALL_ACTIVE_TRAITS,
+    FETCH_AVAILABLE_TRAITS,
+    SET_SELECTED_MAJOR_TRAIT_ID,
+    WIPE_SELECTED_TRAIT_IDS,
+    WIPE_ALL_SELECTED_TRAIT_IDS,
 
-    fetchTraits,
-    setMajorTrait,
-    wipeActiveTraits,
-    wipeAllActiveTraits
+    fetchAvailableTraits,
+    setSelectedMajorTraitId,
+    wipeSelectedTraitIds,
+    wipeAllSelectedTraitIds
 };
