@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 import MechanicBar from './MechanicBar';
 
 const mapStateToProps = state => ({
-    profession: state.profession
+    selectedProfession: state.selectedProfession
 });
 
 export default connect(mapStateToProps)(MechanicBar);
