@@ -5,14 +5,14 @@ import SkillIcon from '../../../../Icon';
 import SkillTooltip from '../../../../../Tooltips/Skills/TooltipContainer';
 import style from './style.css';
 
-const renderSkills = (professionSkills, skills) => {
-    const slots = groupBy(professionSkills.filter(skill => skill.type === 'Profession'), 'slot');
+const renderSkills = (availableProfessionSkillObjects, availableSkillObjects) => {
+    const slots = groupBy(availableProfessionSkillObjects.filter(skill => skill.type === 'Profession'), 'slot');
 
     return ['Profession_1', 'Profession_2', 'Profession_3', 'Profession_4'].map(slot => {
         if (slots[slot]) {
             return (
-                <SkillTooltip skill={skills[slots[slot][0].id]} key={slot}>
-                    <SkillIcon skill={skills[slots[slot][0].id]} size={32}/>
+                <SkillTooltip skill={availableSkillObjects[slots[slot][0].id]} key={slot}>
+                    <SkillIcon skill={availableSkillObjects[slots[slot][0].id]} size={32}/>
                 </SkillTooltip>
             );
         } else {
@@ -21,12 +21,12 @@ const renderSkills = (professionSkills, skills) => {
     });
 };
 
-const Mesmer = ({ professionSkills, skills }) => (
+const Mesmer = ({ availableProfessionSkillObjects, availableSkillObjects }) => (
     <div className={style.component}>
         <div className={style.illusions}>
             {range(0, 3).map(i => (<span key={i}/>))}
         </div>
-        {renderSkills(professionSkills, skills)}
+        {renderSkills(availableProfessionSkillObjects, availableSkillObjects)}
     </div>
 );
 
