@@ -4,9 +4,9 @@ import SkillIcon from '../../../../Icon';
 import SkillTooltip from '../../../../../Tooltips/Skills/TooltipContainer';
 import style from './style.css';
 
-const renderSkill = (professionSkills, skills) => {
-    const possibleSkill = professionSkills.filter(skill => skill.slot === 'Profession_1')[0];
-    const skill = possibleSkill ? skills[possibleSkill.id] : null;
+const renderSkill = (availableProfessionSkillObjects, availableSkillObjects) => {
+    const possibleSkill = availableProfessionSkillObjects.filter(skill => skill.slot === 'Profession_1')[0];
+    const skill = possibleSkill ? availableSkillObjects[possibleSkill.id] : null;
 
     return (
         <SkillTooltip skill={skill}>
@@ -15,9 +15,9 @@ const renderSkill = (professionSkills, skills) => {
     );
 };
 
-const Thief = ({ professionSkills, skills }) => (
+const Thief = ({ availableProfessionSkillObjects, availableSkillObjects }) => (
     <div className={style.component}>
-        {renderSkill(professionSkills, skills)}
+        {renderSkill(availableProfessionSkillObjects, availableSkillObjects)}
         <SkillIcon.Empty size={32}/>
         <div className={style.initiative}>
             {range(0, 15).map(i => (<span key={i}/>))}
