@@ -1,11 +1,11 @@
 import { connect } from 'react-redux';
-import { getActiveMainhand } from '~/selectors/skills/weapons';
+import { getActiveMainhandWeaponId } from '~/selectors/skills/weapons';
 import Warrior from './Component';
 
 const mapStateToProps = (state, ownProps) => ({
-    skills: state.skills,
-    professionSkills: state.professionSkills,
-    weapon: getActiveMainhand(state, ownProps)
+    availableSkillObjects: state.availableSkillObjects,
+    availableProfessionSkillObjects: state.availableProfessionSkillObjects,
+    weapon: getActiveMainhandWeaponId(state, ownProps)
 });
 
 export default connect(mapStateToProps)(Warrior);
