@@ -1,15 +1,15 @@
 import { connect } from 'react-redux';
-import { getActiveMainhand, getActiveOffhand, getIsTwoHandedActive } from '~/selectors/skills';
+import { getActiveMainhandWeaponId, getActiveOffhandWeaponId, getIsTwoHandedActive } from '~/selectors/skills';
 import { getActiveAttunement } from '~/selectors/skills/mechanic';
 import WeaponBar from './WeaponBar';
 
 const mapStateToProps = (state, ownProps) => ({
-    mainhand: getActiveMainhand(state, ownProps),
-    offhand: getActiveOffhand(state, ownProps),
-    isTwoHanded: getIsTwoHandedActive(state, ownProps),
-    attunement: getActiveAttunement(state, ownProps),
-    weapons: state.weapons,
-    skills: state.skills
+    activeMainhandWeaponId: getActiveMainhandWeaponId(state, ownProps),
+    activeOffhandWeaponId: getActiveOffhandWeaponId(state, ownProps),
+    activeAttunement: getActiveAttunement(state, ownProps),
+    availableSkillObjects: state.availableSkillObjects,
+    availableWeaponObjects: state.availableWeaponObjects,
+    isTwoHanded: getIsTwoHandedActive(state, ownProps)
 });
 
 export default connect(mapStateToProps)(WeaponBar);
