@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import { setWeaponSet } from '~/actions';
+import { setActiveWeaponSet } from '~/actions';
 import SetSelection from './SetSelection';
 import { getHasMultipleWeaponsets } from '~/selectors/skills';
 
@@ -10,7 +10,7 @@ const mapStateToProps = state => ({
 
 const mapDispatchToProps = dispatch => ({
     onWeaponSetChange: activeWeaponSet => {
-        dispatch(setWeaponSet({
+        dispatch(setActiveWeaponSet({
             activeWeaponSet
         }));
     }
