@@ -3,9 +3,9 @@ import { getSelectedSpecializationId } from '~/selectors/specializations';
 import SpecializationLine from './SpecializationLine';
 
 const mapStateToProps = (state, ownProps) => ({
-    selectedSpecialization: getSelectedSpecializationId(state, ownProps),
-    specializations: state.specializations,
-    traits: state.traits
+    availableSpecializationObjects: state.availableSpecializationObjects,
+    availableTraitObjects: state.availableTraitObjects,
+    selectedSpecializationId: getSelectedSpecializationId(state, ownProps)
 });
 
 export default connect(mapStateToProps)(SpecializationLine);
