@@ -2,8 +2,8 @@ import { connect } from 'react-redux';
 import TraitTooltip from './Tooltip';
 
 const mapStateToProps = state => ({
-    activeMajorTraits: state.activeMajorTraits,
-    activeMinorTraits: state.activeMinorTraits
+    selectedMajorTraitIds: state.selectedMajorTraitIds,
+    selectedMinorTraitIds: state.selectedMinorTraitIds
 });
 
 export default connect(mapStateToProps)(TraitTooltip);
