@@ -1,22 +1,22 @@
-import * as actions from '../actions';
+import * as actions from '~/actions';
 import { handleSimpleAction } from './utils';
 import { syncedReducer } from 'redux-sync-reducer';
 
 /** The language reducer. */
-export const language = syncedReducer(handleSimpleAction(actions.SET_LANGUAGE, 'en', 'language'), { name: 'language' });
+export const selectedLanguage = syncedReducer(handleSimpleAction(actions.SET_SELECTED_LANGUAGE, 'en', 'language'), { name: 'language' });
 
 /** The game mode reducer. */
-export const gameMode = handleSimpleAction(actions.SET_GAMEMODE, null, 'gameMode');
+export const selectedGameMode = handleSimpleAction(actions.SET_SELECTED_GAMEMODE, null, 'gameMode');
 
 /** The profession reducer. */
-export const profession = handleSimpleAction(actions.SET_PROFESSION, null, 'profession');
+export const selectedProfession = handleSimpleAction(actions.SET_SELECTED_PROFESSION, null, 'profession');
 
 /** The race reducer. */
-export const race = handleSimpleAction(actions.SET_RACE, 'none', 'race');
+export const selectedRace = handleSimpleAction(actions.SET_SELECTED_RACE, null, 'race');
 
 export default {
-    language,
-    gameMode,
-    profession,
-    race
+    selectedLanguage,
+    selectedGameMode,
+    selectedProfession,
+    selectedRace
 };
