@@ -1,7 +1,7 @@
 import { handleAction, handleActions } from 'redux-actions';
-import * as actions from '../../actions';
+import * as actions from '~/actions';
 
-export const attunements = handleAction(actions.FETCH_PROFESSION, (state, action) => {
+export const availableAttunementObjects = handleAction(actions.FETCH_PROFESSION, (state, action) => {
     return action.payload.attunements || {};
 }, {});
 
@@ -9,16 +9,16 @@ export const attunements = handleAction(actions.FETCH_PROFESSION, (state, action
 const DEFAULT_ATTUNEMENTS = ['Fire', 'Fire', 'Fire', 'Fire'];
 
 export const activeAttunements = handleActions({
-    [actions.SET_ATTUNEMENT]: (state, action) => {
+    [actions.SET_ACTIVE_ATTUNEMENT]: (state, action) => {
         const newState = state.slice();
         newState[action.payload.activeWeaponSet] = action.payload.attunement;
         return newState;
     },
-    [actions.WIPE_ALL_WEAPONS]: () => DEFAULT_ATTUNEMENTS,
-    [actions.SET_PROFESSION]: () => DEFAULT_ATTUNEMENTS
+    [actions.WIPE_ALL_SELECTED_WEAPON_IDS]: () => DEFAULT_ATTUNEMENTS,
+    [actions.SET_SELECTED_PROFESSION]: () => DEFAULT_ATTUNEMENTS
 }, DEFAULT_ATTUNEMENTS);
 
 export default {
     activeAttunements,
-    attunements
+    availableAttunementObjects
 };
