@@ -1,10 +1,10 @@
 import forEach from 'lodash/forEach';
 import { handleAction } from 'redux-actions';
-import * as actions from '../../actions';
-import { handleSimpleAction } from '../utils';
+import * as actions from '~/actions';
+import { handleSimpleAction } from '~/reducers/utils';
 
 /** Reducer for the available skill ids for the current profession. */
-export const skillIds = handleAction(actions.FETCH_PROFESSION, (state, action) => {
+export const availableSkillIds = handleAction(actions.FETCH_PROFESSION, (state, action) => {
     const skills = [];
 
     forEach(action.payload.weapons, weapon => {
@@ -22,14 +22,14 @@ export const skillIds = handleAction(actions.FETCH_PROFESSION, (state, action) =
     return skills;
 }, []);
 
-export const professionSkills = handleAction(actions.FETCH_PROFESSION, (state, action) => {
+export const availableSkillObjects = handleSimpleAction(actions.FETCH_AVAILABLE_SKILLS, {});
+
+export const availableProfessionSkillObjects = handleAction(actions.FETCH_PROFESSION, (state, action) => {
     return action.payload.skills;
 }, []);
 
-export const skills = handleSimpleAction(actions.FETCH_SKILLS, {});
-
 export default {
-    skillIds,
-    professionSkills,
-    skills
+    availableSkillIds,
+    availableSkillObjects,
+    availableProfessionSkillObjects
 };
