@@ -1,7 +1,7 @@
 import { serialize, deserialize } from 'gw2be-build-string';
-import { setGameMode, setProfession, setRace } from '../actions/general';
-import { setSpecialization } from '../actions/specializations';
-import { setMajorTrait } from '../actions/traits';
+import { setSelectedGameMode, setSelectedProfession, setSelectedRace } from '../actions/general';
+import { setSelectedSpecializationId } from '../actions/specializations';
+import { setSelectedMajorTraitId } from '../actions/traits';
 
 /**
  * Initializes a build.
@@ -29,27 +29,27 @@ export function initializeBuildFromString(store, buildString) {
     };
 
     return Promise.all([
-        disp('gameMode', build.general, setGameMode),
-        disp('profession', build.general, setProfession),
-        disp('race', build.general, setRace)
+        disp('gameMode', build.general, setSelectedGameMode),
+        disp('profession', build.general, setSelectedProfession),
+        disp('race', build.general, setSelectedRace)
     ]).then(() => {
-        const specializations = store.getState().specializations;
+        const specializations = store.getState().availableSpecializationObjects;
         return Promise.all([
-            disp('specialization', build.specialization1, setSpecialization, id => ({ specializationLine: 0, specializationId: id, specializations })),
-            disp('specialization', build.specialization2, setSpecialization, id => ({ specializationLine: 1, specializationId: id, specializations })),
-            disp('specialization', build.specialization3, setSpecialization, id => ({ specializationLine: 2, specializationId: id, specializations }))
+            disp('specialization', build.specialization1, setSelectedSpecializationId, id => ({ specializationLine: 0, specializationId: id, specializations })),
+            disp('specialization', build.specialization2, setSelectedSpecializationId, id => ({ specializationLine: 1, specializationId: id, specializations })),
+            disp('specialization', build.specialization3, setSelectedSpecializationId, id => ({ specializationLine: 2, specializationId: id, specializations }))
         ]);
     }).then(() => {
         return Promise.all([
-            disp('majorTrait1', build.specialization1, setMajorTrait, id => ({ specializationLine: 0, traitTier: 1, traitId: id })),
-            disp('majorTrait2', build.specialization1, setMajorTrait, id => ({ specializationLine: 0, traitTier: 2, traitId: id })),
-            disp('majorTrait3', build.specialization1, setMajorTrait, id => ({ specializationLine: 0, traitTier: 3, traitId: id })),
-            disp('majorTrait1', build.specialization2, setMajorTrait, id => ({ specializationLine: 1, traitTier: 1, traitId: id })),
-            disp('majorTrait2', build.specialization2, setMajorTrait, id => ({ specializationLine: 1, traitTier: 2, traitId: id })),
-            disp('majorTrait3', build.specialization2, setMajorTrait, id => ({ specializationLine: 1, traitTier: 3, traitId: id })),
-            disp('majorTrait1', build.specialization3, setMajorTrait, id => ({ specializationLine: 2, traitTier: 1, traitId: id })),
-            disp('majorTrait2', build.specialization3, setMajorTrait, id => ({ specializationLine: 2, traitTier: 2, traitId: id })),
-            disp('majorTrait3', build.specialization3, setMajorTrait, id => ({ specializationLine: 2, traitTier: 3, traitId: id }))
+            disp('majorTrait1', build.specialization1, setSelectedMajorTraitId, id => ({ specializationLine: 0, traitTier: 1, traitId: id })),
+            disp('majorTrait2', build.specialization1, setSelectedMajorTraitId, id => ({ specializationLine: 0, traitTier: 2, traitId: id })),
+            disp('majorTrait3', build.specialization1, setSelectedMajorTraitId, id => ({ specializationLine: 0, traitTier: 3, traitId: id })),
+            disp('majorTrait1', build.specialization2, setSelectedMajorTraitId, id => ({ specializationLine: 1, traitTier: 1, traitId: id })),
+            disp('majorTrait2', build.specialization2, setSelectedMajorTraitId, id => ({ specializationLine: 1, traitTier: 2, traitId: id })),
+            disp('majorTrait3', build.specialization2, setSelectedMajorTraitId, id => ({ specializationLine: 1, traitTier: 3, traitId: id })),
+            disp('majorTrait1', build.specialization3, setSelectedMajorTraitId, id => ({ specializationLine: 2, traitTier: 1, traitId: id })),
+            disp('majorTrait2', build.specialization3, setSelectedMajorTraitId, id => ({ specializationLine: 2, traitTier: 2, traitId: id })),
+            disp('majorTrait3', build.specialization3, setSelectedMajorTraitId, id => ({ specializationLine: 2, traitTier: 3, traitId: id }))
         ]);
     });
 }
@@ -62,27 +62,27 @@ export function initializeBuildFromString(store, buildString) {
 export function exportBuildToString(state) {
     const build = {
         general: {
-            gameMode: state.gameMode,
-            profession: state.profession,
-            race: state.race
+            gameMode: state.selectedGameMode,
+            profession: state.selectedProfession,
+            race: state.selectedRace
         },
         specialization1: {
-            specialization: state.activeSpecializations[0],
-            majorTrait1: state.activeMajorTraits[0],
-            majorTrait2: state.activeMajorTraits[1],
-            majorTrait3: state.activeMajorTraits[2]
+            specialization: state.selectedSpecializationIds[0],
+            majorTrait1: state.selectedMajorTraitIds[0],
+            majorTrait2: state.selectedMajorTraitIds[1],
+            majorTrait3: state.selectedMajorTraitIds[2]
         },
         specialization2: {
-            specialization: state.activeSpecializations[1],
-            majorTrait1: state.activeMajorTraits[3],
-            majorTrait2: state.activeMajorTraits[4],
-            majorTrait3: state.activeMajorTraits[5]
+            specialization: state.selectedSpecializationIds[1],
+            majorTrait1: state.selectedMajorTraitIds[3],
+            majorTrait2: state.selectedMajorTraitIds[4],
+            majorTrait3: state.selectedMajorTraitIds[5]
         },
         specialization3: {
-            specialization: state.activeSpecializations[2],
-            majorTrait1: state.activeMajorTraits[6],
-            majorTrait2: state.activeMajorTraits[7],
-            majorTrait3: state.activeMajorTraits[8]
+            specialization: state.selectedSpecializationIds[2],
+            majorTrait1: state.selectedMajorTraitIds[6],
+            majorTrait2: state.selectedMajorTraitIds[7],
+            majorTrait3: state.selectedMajorTraitIds[8]
         }
     };
 
