@@ -4,7 +4,7 @@ import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
 import {
-    skills, weapons, activeMajorTraits, activeMinorTraits,
+    availableSkillObjects, availableWeaponObjects, selectedMajorTraits, selectedMinorTraits,
     WEAPON_SIMPLE, WEAPON_OFFHAND, WEAPON_ATTUNEMENT,
     ATTUNEMENT1, ATTUNEMENT2,
     SKILL_SIMPLE, SKILL_OFFHAND_NOTHING, SKILL_OFFHAND_SOMETHING, SKILL_ATTUNEMENT_1, SKILL_ATTUNEMENT_2
@@ -15,14 +15,14 @@ import SkillIcon from '~/components/Skills/Icon';
 
 describe('WeaponBar', () => {
     it('renders 5 empty skill icons', () => {
-        const props = { skills, weapons, activeMajorTraits, activeMinorTraits };
+        const props = { availableSkillObjects, availableWeaponObjects, selectedMajorTraits, selectedMinorTraits };
         expect(shallow(<WeaponBar {...props}/>).find(SkillIcon.Empty)).to.have.length(5);
     });
 
     it('renders mainhand skills', () => {
         const props = {
-            skills, weapons, activeMajorTraits, activeMinorTraits,
-            mainhand: WEAPON_SIMPLE, isTwoHanded: false
+            availableSkillObjects, availableWeaponObjects, selectedMajorTraits, selectedMinorTraits,
+            activeMainhandWeaponId: WEAPON_SIMPLE, isTwoHanded: false
         };
 
         const wrapper = shallow(<WeaponBar {...props}/>);
@@ -33,8 +33,8 @@ describe('WeaponBar', () => {
 
     it('renders offhand skills', () => {
         const props = {
-            skills, weapons, activeMajorTraits, activeMinorTraits,
-            offhand: WEAPON_SIMPLE, isTwoHanded: false
+            availableSkillObjects, availableWeaponObjects, selectedMajorTraits, selectedMinorTraits,
+            activeOffhandWeaponId: WEAPON_SIMPLE, isTwoHanded: false
         };
 
         const wrapper = shallow(<WeaponBar {...props}/>);
@@ -44,8 +44,8 @@ describe('WeaponBar', () => {
 
     it('renders mainhand and offhand skills', () => {
         const props = {
-            skills, weapons, activeMajorTraits, activeMinorTraits,
-            mainhand: WEAPON_SIMPLE, offhand: WEAPON_SIMPLE, isTwoHanded: false
+            availableSkillObjects, availableWeaponObjects, selectedMajorTraits, selectedMinorTraits,
+            activeMainhandWeaponId: WEAPON_SIMPLE, activeOffhandWeaponId: WEAPON_SIMPLE, isTwoHanded: false
         };
 
         const wrapper = shallow(<WeaponBar {...props}/>);
@@ -55,8 +55,8 @@ describe('WeaponBar', () => {
 
     it('renders two handed skills', () => {
         const props = {
-            skills, weapons, activeMajorTraits, activeMinorTraits,
-            mainhand: WEAPON_SIMPLE, isTwoHanded: true
+            availableSkillObjects, availableWeaponObjects, selectedMajorTraits, selectedMinorTraits,
+            activeMainhandWeaponId: WEAPON_SIMPLE, isTwoHanded: true
         };
 
         const wrapper = shallow(<WeaponBar {...props}/>);
@@ -66,25 +66,25 @@ describe('WeaponBar', () => {
 
     it('renders offhand dependant skills', () => {
         const props = {
-            skills, weapons, activeMajorTraits, activeMinorTraits,
-            mainhand: WEAPON_OFFHAND, isTwoHanded: false
+            availableSkillObjects, availableWeaponObjects, selectedMajorTraits, selectedMinorTraits,
+            activeMainhandWeaponId: WEAPON_OFFHAND, isTwoHanded: false
         };
 
         expect(shallow(<WeaponBar {...props}/>)
             .find(SkillIcon).at(2).prop('skill')).to.equal(SKILL_OFFHAND_NOTHING);
-        expect(shallow(<WeaponBar {...props} offhand={WEAPON_SIMPLE}/>)
+        expect(shallow(<WeaponBar {...props} activeOffhandWeaponId={WEAPON_SIMPLE}/>)
             .find(SkillIcon).at(2).prop('skill')).to.equal(SKILL_OFFHAND_SOMETHING);
     });
 
     it('renders attunement dependant skills', () => {
         const props = {
-            skills, weapons, activeMajorTraits, activeMinorTraits,
-            mainhand: WEAPON_ATTUNEMENT, isTwoHanded: true
+            availableSkillObjects, availableWeaponObjects, selectedMajorTraits, selectedMinorTraits,
+            activeMainhandWeaponId: WEAPON_ATTUNEMENT, isTwoHanded: true
         };
 
-        expect(shallow(<WeaponBar {...props} attunement={ATTUNEMENT1}/>)
+        expect(shallow(<WeaponBar {...props} activeAttunement={ATTUNEMENT1}/>)
             .find(SkillIcon).first().prop('skill')).to.equal(SKILL_ATTUNEMENT_1);
-        expect(shallow(<WeaponBar {...props} attunement={ATTUNEMENT2}/>)
+        expect(shallow(<WeaponBar {...props} activeAttunement={ATTUNEMENT2}/>)
             .find(SkillIcon).first().prop('skill')).to.equal(SKILL_ATTUNEMENT_2);
     });
 });
